requirements.txt
python-dotenv
uagents
langchain
openai  
flask
flask-cors
requests
sqlalchemy
websocket-client
pytest
pytest-mock
flake8

# React Dashboard build tools (for completeness)
# Note: Run `npm install` in the dashboard directory for these dependencies.
